+++ adapters/DTExchange/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## DT Exchange iOS Mediation Adapter Changelog
 
+#### Next version
+- Implemented DTExchange SDK Native Ads Adapter.
+
 #### [Version 8.4.2.0](https://dl.google.com/googleadmobadssdk/mediation/ios/dtexchange/DTExchangeAdapter-8.4.2.0.zip)
 - Verified compatibility with DT Exchange SDK version 8.4.2.